feat: create instructions
