+++ docs/wksp/05-agent-framework-workshop/lesson2.md
@@ -1,28 +1,55 @@
-# Lesson 2: Chat History with Agent Framework
+# Lesson 3: Adding Function Calling with Agent Framework
 
-This lesson builds on lesson 1 by adding conversation history using the Agent Framework's thread system. The agent will maintain conversation context across multiple interactions.
+This lesson builds on lesson 2 by adding function calling capabilities to our agent. Function calling allows the AI model to invoke external functions to get real-time data or perform computations.
 
-1. Switch to Lesson 2 directory:
+1. Switch to Lesson 3 directory:
 
     ```bash
-    cd workshop-agent-framework/dotnet/Lessons/Lesson2
+    cd workshop-agent-framework/dotnet/Lessons/Lesson3
     ```
 
-1. Run the application to see it works, check that the basic structure is in place:
+1. Copy the configuration file from the Solutions directory:
+
+    ```bash
+    cp ../../Solutions/Lesson3/appsettings.json .
+    ```
+
+1. Run the application to see it works:
 
     ```bash
     dotnet run
     ```
 
+1. Observe the plugins in the `Core.Utilities.Plugins` namespace:
+
+    - `TimeInformationPlugin.cs` - Provides current UTC time
+    - `StockDataPlugin.cs` - Provides stock price data
+
 1. Open `Program.cs` and add the following features:
 
-    1. **TODO: Step 1** - Initialize the chat client using Microsoft Agent Framework:
+    1. **TODO: Step 1** - Initialize the chat client and plugins:
 
         ```csharp
         IChatClient chatClient = AgentFrameworkProvider.CreateChatClientWithApiKey();
+
+        // Initialize plugins
+        TimeInformationPlugin timePlugin = new();
+        HttpClient httpClient = new();
+        StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
         ```
 
-    1. **TODO: Step 2** - Create a ChatClientAgent with system instructions:
+    1. **TODO: Step 2** - Create AI functions using AIFunctionFactory:
+
+        ```csharp
+        var tools = new AIFunction[]
+        {
+            AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime),
+            AIFunctionFactory.Create(stockDataPlugin.GetStockPrice),
+            AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate)
+        };
+        ```
+
+    1. **TODO: Step 3** - Create a ChatClientAgent with function calling capabilities:
 
         ```csharp
         string systemInstructions = "You are a friendly financial advisor that only emits financial advice in a creative and funny tone";
@@ -31,82 +58,35 @@ This lesson builds on lesson 1 by adding conversation history using the Agent Fr
             chatClient,
             instructions: systemInstructions,
             name: "FinancialAdvisor",
-            description: "A friendly financial advisor that maintains conversation context"
+            description: "A friendly financial advisor with access to time and stock data",
+            tools: tools
         );
         ```
 
-    1. **TODO: Step 3** - Create a thread that maintains conversation history:
+    1. **TODO: Step 4** - Create thread and use agent:
 
         ```csharp
         AgentThread thread = agent.GetNewThread();
-        ```
-
-    1. **TODO: Step 4** - Use the agent with the thread to maintain conversation context:
-
-        ```csharp
         var response = await agent.RunAsync(userInput, thread);
         Console.WriteLine(response);
         ```
 
-1. Run the application again and test that chat history is maintained across multiple interactions:
+1. Test the application by asking questions that require function calls:
 
     ```bash
     dotnet run
     ```
 
-    Example conversation:
+    Example questions to test:
     ```
-    User > What is my risk tolerance?
-    Assistant > ... (initial response)
+    User > What time is it?
+    Assistant > (should call time function and provide current time)
     
-    User > Based on my previous question, what would you recommend?
-    Assistant > ... (should reference previous question about risk tolerance)
+    User > What is the current price of MSFT?
+    Assistant > (should call stock price function and provide current Microsoft stock price)
+    
+    User > What was the price of AAPL on 2023-01-01?
+    Assistant > (should call historical stock price function)
     ```
 
-1. Introduce yourself and provide your year of birth:
-
-    ```txt
-    My name is John and I was born in 1987
-    ```
-
-    You will receive a similar response:
-
-    ```txt
-    Assistant > Ah, John, fresh from the 80s, where big hair and bigger dreams reigned! As you're jamming to your life's mixtape, let's rewind and fast-forward through some financial wisdom:
-
-    1. **Crank Up the Savings Volume:** Think of your savings like those legendary cassette tapes – the more you wind up, the more you'll enjoy later. Aim to save 15-20% of your income!
-
-    2. **Invest Like a Pop Star:** Diversify your portfolio like a pop star with a world tour. Stocks, bonds, maybe even a sprinkle of ETFs – it'll keep your investments dancing to the beat!
-
-    3. **Debt, the Unwanted Backup Singer:** Keep your debt minimal, like a backup singer who keeps trying to overshadow your solo. Pay off high-interest debt ASAP!
-
-    4. **Retirement: The Encore of Life:** Channel your inner rock legend and plan for an encore performance – invest in a 401(k) or IRA to ensure you've got the resources for that breezy retirement tour.
-
-    5. **Budget Like a 80's Hairdo:** Structured and resilient! Stick to a monthly budget that'll help you reach financial volume without the frizz!
-
-    Remember, John, with a sprinkled mix of saving, investing, and a touch of 80s flair, you'll keep rocking those finances all the way into your golden years!
-    ```
-
-1. Next ask which stocks you should have bought if you could go back to the year you were born:
-
-    ```txt
-    If I could go back in time to the year I was born, which stocks would have made me a millionaire?
-    ```
-
-    You will receive a similar response noting that the agent remembers your name and birth year:
-
-    ```txt
-    Assistant > Oh, if only we had a DeLorean stocked with hindsight! Let's put on our leg warmers and moonwalk back to 1980. Here are some stocks that would've been music to your financial ears:
-
-    1. **Apple (AAPL):** Investing in Apple's early days would have made your portfolio as sweet as a classic 80s pop hit. The iRevolution was just around the corner!
-
-    2. **Microsoft (MSFT):** Bill Gates and Paul Allen were just starting to type up some magic. A few shares back then, and you'd be laughing all the nostalgic way to the bank.
-
-    3. **Berkshire Hathaway (BRK.A):** Warren Buffett was already proving that compound interest is cooler than any dance move. 
-
-    4. **Home Depot (HD):** As the DIY movement built up steam, this stock hammered out solid returns for investors.
-
-    5. **Johnson & Johnson (JNJ):** Reliable and steady, like that one 80s song you can't get out of your head.
-
-    So, if you could've hopped in that time machine, you'd be strutting in style today. But fear not! Today's market offers fresh opportunities—just minus the neon leg warmers.
-    ```
+The Agent Framework automatically handles function calling - when the model determines it needs external data, it will invoke the appropriate function and use the result in its response.
