+++ docs/wksp/05-agent-framework-workshop/lesson3.md
@@ -1,11 +1,17 @@
-# Lesson 3: Adding Function Calling with Agent Framework
+# Lesson 4: Specialized Agent with Stock Sentiment Analysis
 
-This lesson builds on lesson 2 by adding function calling capabilities to our agent. Function calling allows the AI model to invoke external functions to get real-time data or perform computations.
+This lesson creates a specialized agent focused on stock sentiment analysis using the Agent Framework with specific system instructions and function calling capabilities.
 
-1. Switch to Lesson 3 directory:
+1. Switch to Lesson 4 directory:
 
     ```bash
-    cd workshop-agent-framework/dotnet/Lessons/Lesson3
+    cd workshop-agent-framework/dotnet/Lessons/Lesson4
+    ```
+
+1. Copy the configuration file from the Solutions directory:
+
+    ```bash
+    cp ../../Solutions/Lesson4/appsettings.json .
     ```
 
 1. Run the application to see it works:
@@ -14,12 +20,7 @@ This lesson builds on lesson 2 by adding function calling capabilities to our ag
     dotnet run
     ```
 
-1. Observe the plugins in the `Core.Utilities.Plugins` namespace:
-
-    - `TimeInformationPlugin.cs` - Provides current UTC time
-    - `StockDataPlugin.cs` - Provides stock price data
-
-1. Open `Program.cs` and add the following features:
+1. Open `Program.cs` and build a specialized stock sentiment agent:
 
     1. **TODO: Step 1** - Initialize the chat client and plugins:
 
@@ -32,40 +33,60 @@ This lesson builds on lesson 2 by adding function calling capabilities to our ag
         StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
         ```
 
-    1. **TODO: Step 2** - Create AI functions using AIFunctionFactory:
+    1. **TODO: Step 2** - Create AI Functions from plugins:
 
         ```csharp
-        var tools = new AIFunction[]
-        {
-            AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime),
-            AIFunctionFactory.Create(stockDataPlugin.GetStockPrice),
-            AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate)
-        };
+        var timeTool = AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime);
+        var stockPriceTool = AIFunctionFactory.Create(stockDataPlugin.GetStockPrice);
+        var stockPriceDateTool = AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate);
         ```
 
-    1. **TODO: Step 3** - Create a ChatClientAgent with function calling capabilities:
+    1. **TODO: Step 3** - Define specialized system instructions for stock sentiment analysis:
 
         ```csharp
-        string systemInstructions = "You are a friendly financial advisor that only emits financial advice in a creative and funny tone";
+        string stockSentimentAgentInstructions = """
+            You are a Stock Sentiment Agent. Your responsibility is to find the stock sentiment for a given Stock.
 
-        ChatClientAgent agent = new(
+            RULES:
+            - Use stock sentiment scale from 1 to 10 where stock sentiment is 1 for sell and 10 for buy.
+            - Provide the rating in your response and a recommendation to buy, hold or sell.
+            - Include the reasoning behind your recommendation.
+            - Include the source of the sentiment in your response.
+            - Focus on technical analysis based on stock price data and general market knowledge.
+            """;
+        ```
+
+    1. **TODO: Step 4** - Create the specialized Stock Sentiment Agent:
+
+        ```csharp
+        ChatClientAgent stockSentimentAgent = new(
             chatClient,
-            instructions: systemInstructions,
-            name: "FinancialAdvisor",
-            description: "A friendly financial advisor with access to time and stock data",
-            tools: tools
+            instructions: stockSentimentAgentInstructions,
+            name: "StockSentimentAgent",
+            description: "An intelligent agent that analyzes stock sentiment using market data",
+            tools: [
+                timeTool,
+                stockPriceTool, 
+                stockPriceDateTool
+            ]
         );
         ```
 
-    1. **TODO: Step 4** - Create thread and use agent:
+    1. **TODO: Step 5** - Create thread and process user requests:
 
         ```csharp
-        AgentThread thread = agent.GetNewThread();
-        var response = await agent.RunAsync(userInput, thread);
-        Console.WriteLine(response);
+        AgentThread thread = stockSentimentAgent.GetNewThread();
+        
+        var response = await stockSentimentAgent.RunAsync(userInput, thread);
+        
+        if (response?.Messages?.Any() == true)
+        {
+            var lastMessage = response.Messages.Last();
+            Console.WriteLine(lastMessage.Text ?? "No response generated.");
+        }
         ```
 
-1. Test the application by asking questions that require function calls:
+1. Test the specialized agent with stock sentiment queries:
 
     ```bash
     dotnet run
@@ -73,14 +94,14 @@ This lesson builds on lesson 2 by adding function calling capabilities to our ag
 
     Example questions to test:
     ```
-    User > What time is it?
-    Assistant > (should call time function and provide current time)
+    User > What is your sentiment on MSFT?
+    Assistant > (should analyze Microsoft stock and provide sentiment rating 1-10 with buy/hold/sell recommendation)
     
-    User > What is the current price of MSFT?
-    Assistant > (should call stock price function and provide current Microsoft stock price)
+    User > Analyze AAPL sentiment
+    Assistant > (should provide Apple stock sentiment analysis with reasoning)
     
-    User > What was the price of AAPL on 2023-01-01?
-    Assistant > (should call historical stock price function)
+    User > Should I buy TSLA?
+    Assistant > (should analyze Tesla and provide specific buy/hold/sell recommendation)
     ```
 
-The Agent Framework automatically handles function calling - when the model determines it needs external data, it will invoke the appropriate function and use the result in its response.
+This lesson demonstrates how to create specialized agents with focused system instructions and domain-specific capabilities using the Agent Framework.
