+++ workshop-agent-framework/dotnet/Lessons/Lesson2/Lesson3.csproj
