+++ workshop-agent-framework/dotnet/Lessons/Lesson2/Program.cs
@@ -1,29 +1,33 @@
 using Core.Utilities.Config;
 using Microsoft.Extensions.AI;
 using Microsoft.Agents.AI;
+// TODO: Step 1 - Add the required using statements for plugins
 
-// TODO: Step 1 - Initialize the chat client using Microsoft Agent Framework
-// (This should be the same as Lesson 1)
+// Step 1 - Initialize the chat client using Microsoft Agent Framework
+IChatClient chatClient = AgentFrameworkProvider.CreateChatClientWithApiKey();
 
+// TODO: Step 2 - Initialize plugins for function calling
 
-// TODO: Step 2 - Create financial advisor agent that maintains conversation context
-// Hint: Use the same ChatClientAgent setup as Lesson 1, but note that Agent Framework
-//       automatically maintains conversation history through the AgentThread
+// TODO: Step 3 - Create AI Functions from plugins
 
+// Step 2 - Configure system message using Agent Framework pattern
+string systemInstructions = "You are a friendly financial advisor that only emits financial advice in a creative and funny tone";
 
-// TODO: Step 3 - Create a thread for conversation with history
-// Hint: Use agent.GetNewThread() - this thread will automatically maintain conversation context
+// Create a simple financial advisor agent using ChatClientAgent
+ChatClientAgent agent = new(
+    chatClient,
+    instructions: systemInstructions,
+    name: "FinancialAdvisor",
+    description: "A friendly financial advisor that maintains conversation context"
+    // TODO: Step 4 - Add tools parameter to enable function calling
+);
 
+// Create a thread for conversation
+AgentThread thread = agent.GetNewThread();
 
-// Execute program with conversation history demonstration
+// Execute program.
 const string terminationPhrase = "quit";
 string? userInput;
-
-Console.WriteLine("=== Financial Advisor with Conversation Memory ===");
-Console.WriteLine("This lesson demonstrates how Agent Framework automatically maintains conversation history.");
-Console.WriteLine("Try asking follow-up questions to see the agent remember previous context!");
-Console.WriteLine("Type 'quit' to exit.");
-Console.WriteLine();
 
 do
 {
@@ -41,14 +45,9 @@ do
     {
         Console.Write("Assistant > ");
         
-        // TODO: Step 4 - Use agent to respond with conversation history maintained automatically
-        // Hint: Same as Lesson 1, but notice how the agent remembers previous messages
-        //       The AgentThread automatically maintains this context for you
-        // var response = await agent.RunAsync(userInput, thread);
-        // Console.WriteLine(response);
-
+        // Step 3 - Use agent to respond to user input
+        var response = await agent.RunAsync(userInput, thread);
+        Console.WriteLine(response);
     }
 }
 while (userInput != terminationPhrase);
-
-Console.WriteLine("Notice how the agent remembered our conversation context! This is automatic with Agent Framework.");
