+++ workshop-agent-framework/dotnet/Lessons/Lesson2/README.md
@@ -1,53 +1,78 @@
-# Lesson 2: Conversation Memory with Agent Framework
+# Lesson 3: Function Calling with Agents
 
 ## Overview
-This lesson demonstrates how the Microsoft Agent Framework automatically handles conversation memory and context. Unlike traditional chat implementations where you manually manage chat history, Agent Framework's AgentThread automatically maintains conversation context for you.
+This lesson introduces function calling capabilities in Agent Framework. You'll learn how to give your AI agent access to real-time data through functions, enabling it to provide current stock prices, time information, and historical data.
 
 ## Learning Objectives
-- Understand how AgentThread maintains conversation history automatically
-- Experience the difference between stateless and stateful conversations
-- Learn about the built-in conversation management in Agent Framework
+- Understand how to add function calling capabilities to agents
+- Learn to use AIFunctionFactory to convert methods into AI functions
+- Experience automatic function calling in conversational AI
+- Work with real-time data in AI applications
 
 ## Key Concepts
-- **Automatic Context Management**: Agent Framework handles conversation history behind the scenes
-- **AgentThread Persistence**: The thread object maintains state across multiple interactions
-- **Stateful vs Stateless**: Comparison with manual chat history management approaches
+- **AIFunction**: Wrapper that makes regular methods callable by AI agents
+- **AIFunctionFactory**: Utility for converting methods to AIFunction objects
+- **Automatic Function Calling**: Agent Framework automatically determines when to call functions
+- **Plugin Architecture**: Organized way to group related functions
 
-## What's Different from Lesson 1?
-While the code looks very similar to Lesson 1, the key difference is understanding that:
-- The `AgentThread` automatically stores all conversation history
-- Each call to `agent.RunAsync()` has access to the full conversation context
-- No manual management of `ChatHistory` is required
+## New Components
+
+### Plugins Used
+- **TimeInformationPlugin**: Provides current time information
+- **StockDataPlugin**: Provides current and historical stock price data
+- **StocksService**: Backend service for retrieving stock market data
+
+### Function Conversion
+```csharp
+AIFunctionFactory.Create(plugin.Method)
+```
+This converts a regular C# method into an AIFunction that the agent can call automatically.
 
 ## Steps to Complete
 
-### Step 1: Initialize the Chat Client
-Same as Lesson 1 - create an `IChatClient` using `AgentFrameworkProvider.CreateChatClientWithApiKey()`.
+### Step 1: Initialize Chat Client
+Same as previous lessons.
 
-### Step 2: Create the ChatClientAgent
-Same setup as Lesson 1, but with awareness that this agent will maintain conversation context.
+### Step 2: Initialize Plugins
+Create instances of the plugins that provide the functionality:
+- TimeInformationPlugin for time-related queries
+- StockDataPlugin with StocksService for market data
 
-### Step 3: Create the AgentThread
-Use `agent.GetNewThread()` - this thread will automatically track all conversation history.
+### Step 3: Create Tools Array
+Convert plugin methods to AIFunction objects using AIFunctionFactory.Create():
+- Time function from timePlugin.GetCurrentUtcTime
+- Stock price function from stockDataPlugin.GetStockPrice
+- Historical data function from stockDataPlugin.GetStockPriceForDate
 
-### Step 4: Process Conversations with Memory
-Use `agent.RunAsync(userInput, thread)` and observe how the agent remembers previous exchanges.
+### Step 4: Create Agent with Tools
+Create ChatClientAgent and pass the tools array to enable function calling.
 
-## Testing Conversation Memory
-Try these conversation patterns to see memory in action:
+### Step 5: Create Thread
+Same as previous lessons.
 
-1. **Reference Test**: 
-   - "What's a good investment strategy?"
-   - "Why did you recommend that?" (tests if agent remembers the previous recommendation)
+### Step 6: Use Agent with Function Calling
+The agent will automatically call functions when needed to answer user questions.
 
-2. **Context Building**:
-   - "I'm 25 years old"
-   - "I have $10,000 to invest"
-   - "What should I do?" (tests if agent remembers your age and amount)
+## Testing Function Calling
 
-3. **Follow-up Questions**:
-   - Ask about stocks
-   - "What about the risks?" (tests contextual understanding)
+Try these queries to see automatic function calling:
+
+### Time Queries
+- "What time is it?"
+- "What's the current UTC time?"
+
+### Current Stock Prices
+- "What's the current price of MSFT?"
+- "How is Apple stock doing today?"
+- "Get me Tesla's current stock price"
+
+### Historical Data
+- "What was AAPL's price on 2023-01-15?"
+- "Show me Microsoft's stock price from last month"
+
+### Combined Queries
+- "What time is it and what's Google's current stock price?"
+- "Give me a financial report with current time and top tech stock prices"
 
 ## Expected Behavior
-The agent should demonstrate clear memory of previous exchanges, referencing earlier parts of the conversation and building upon established context. This happens automatically without any additional code for history management.
+The agent should automatically determine when function calls are needed and execute them seamlessly within the conversation. You'll see the agent providing real-time, accurate data while maintaining its creative and funny personality.
