+++ workshop-agent-framework/dotnet/Lessons/Lesson3/Lesson4.csproj
