+++ workshop-agent-framework/dotnet/Lessons/Lesson3/Program.cs
@@ -3,47 +3,49 @@ using Core.Utilities.Plugins;
 using Core.Utilities.Services;
 using Microsoft.Extensions.AI;
 using Microsoft.Agents.AI;
+// TODO: Step 1 - Add the Extensions namespace for HostedWebSearchTool
 
-// TODO: Step 1 - Initialize the chat client with Agent Framework  
+// Initialize the chat client with Agent Framework  
+IChatClient chatClient = AgentFrameworkProvider.CreateChatClientWithApiKey();
 
+// Initialize plugins
+TimeInformationPlugin timePlugin = new();
+HttpClient httpClient = new();
+StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
 
-// TODO: Step 2 - Initialize plugins for function calling
-// Hint: Create instances of TimeInformationPlugin, HttpClient, and StockDataPlugin
-// Example:
-// TimeInformationPlugin timePlugin = new();
-// HttpClient httpClient = new();
-// StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
+// Create AI Functions from plugins
+var tools = new AIFunction[]
+{
+    AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime),
+    AIFunctionFactory.Create(stockDataPlugin.GetStockPrice),
+    AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate)
+};
 
+// TODO: Step 2 - Initialize web search tool for sentiment analysis
 
-// TODO: Step 3 - Create tools array for the agent
-// Hint: Use AIFunctionFactory.Create() to convert plugin methods to AIFunction objects
-// You'll need:
-// - timePlugin.GetCurrentUtcTime
-// - stockDataPlugin.GetStockPrice  
-// - stockDataPlugin.GetStockPriceForDate
+// TODO: Step 3 - Create individual AI function variables for clarity
 
+// Create financial advisor agent with function calling capabilities
+string systemInstructions = "You are a friendly financial advisor that only emits financial advice in a creative and funny tone";
 
-// TODO: Step 4 - Create financial advisor agent with function calling capabilities
-// Hint: Same as previous lessons, but add the tools parameter:
-// ChatClientAgent agent = new(chatClient, instructions: systemInstructions, name: "...", description: "...", tools: tools);
+// TODO: Step 4 - Transform into a Stock Sentiment Agent with specialized instructions
 
+ChatClientAgent agent = new(
+    chatClient,
+    instructions: systemInstructions,
+    name: "FinancialAdvisor",
+    description: "A friendly financial advisor with access to time and stock data",
+    tools: tools
+);
 
-// TODO: Step 5 - Create a thread for conversation
+// TODO: Step 5 - Update agent creation to use new instructions and tools
 
+// Create a thread for conversation
+AgentThread thread = agent.GetNewThread();
 
-// Execute program with function calling demonstration
+// Execute program
 const string terminationPhrase = "quit";
 string? userInput;
-
-Console.WriteLine("=== Financial Advisor with Function Calling ===");
-Console.WriteLine("This lesson demonstrates how agents can call functions to get real-time data.");
-Console.WriteLine("Try asking about:");
-Console.WriteLine("  - Current time: 'What time is it?'");
-Console.WriteLine("  - Stock prices: 'What's the current price of MSFT?'");
-Console.WriteLine("  - Historical data: 'What was AAPL's price on 2023-01-15?'");
-Console.WriteLine("Type 'quit' to exit.");
-Console.WriteLine();
-
 do
 {
     Console.Write("User > ");
@@ -60,14 +62,9 @@ do
     {
         Console.Write("Assistant > ");
         
-        // TODO: Step 6 - Use agent with automatic function calling
-        // Hint: Same as previous lessons - agent.RunAsync(userInput, thread)
-        // The agent will automatically call functions when needed!
-        // var response = await agent.RunAsync(userInput, thread);
-        // Console.WriteLine(response);
-
+        // Use agent with automatic function calling
+        var response = await agent.RunAsync(userInput, thread);
+        Console.WriteLine(response);
     }
 }
 while (userInput != terminationPhrase);
-
-Console.WriteLine("Notice how the agent automatically called functions to get real-time data!");
