+++ workshop-agent-framework/dotnet/Lessons/Lesson3/README.md
@@ -1,32 +1,37 @@
-# Lesson 3: Function Calling with Agents
+# Lesson 4: Specialized Agent - Stock Sentiment Analysis
 
 ## Overview
-This lesson introduces function calling capabilities in Agent Framework. You'll learn how to give your AI agent access to real-time data through functions, enabling it to provide current stock prices, time information, and historical data.
+This lesson demonstrates how to create a specialized AI agent with domain-specific instructions and behavior. You'll build a Stock Sentiment Agent that analyzes stocks using a standardized scoring system and provides structured investment recommendations.
 
 ## Learning Objectives
-- Understand how to add function calling capabilities to agents
-- Learn to use AIFunctionFactory to convert methods into AI functions
-- Experience automatic function calling in conversational AI
-- Work with real-time data in AI applications
+- Create specialized agents with domain-specific instructions
+- Implement structured response formats and scoring systems
+- Combine function calling with specialized analysis
+- Design professional-grade AI agent behavior
 
 ## Key Concepts
-- **AIFunction**: Wrapper that makes regular methods callable by AI agents
-- **AIFunctionFactory**: Utility for converting methods to AIFunction objects
-- **Automatic Function Calling**: Agent Framework automatically determines when to call functions
-- **Plugin Architecture**: Organized way to group related functions
+- **Specialized Instructions**: Detailed system prompts that define agent behavior
+- **Structured Responses**: Consistent output formats for professional applications
+- **Domain Expertise**: Agent designed for specific use cases
+- **Scoring Systems**: Standardized metrics for analysis (1-10 sentiment scale)
 
-## New Components
+## Agent Specifications
 
-### Plugins Used
-- **TimeInformationPlugin**: Provides current time information
-- **StockDataPlugin**: Provides current and historical stock price data
-- **StocksService**: Backend service for retrieving stock market data
+### Stock Sentiment Agent Rules
+- Uses stock sentiment scale from 1 to 10 (1=sell, 10=buy)
+- Provides rating, recommendation (buy/hold/sell), and reasoning
+- Includes sources for sentiment analysis
+- Focuses on technical analysis based on stock price data
+- Maintains professional, analytical tone
 
-### Function Conversion
-```csharp
-AIFunctionFactory.Create(plugin.Method)
+### Expected Response Format
 ```
-This converts a regular C# method into an AIFunction that the agent can call automatically.
+Stock: [SYMBOL]
+Sentiment Score: [1-10]
+Recommendation: [BUY/HOLD/SELL]
+Reasoning: [Technical analysis explanation]
+Source: [Market data, technical indicators, etc.]
+```
 
 ## Steps to Complete
 
@@ -34,45 +39,48 @@ This converts a regular C# method into an AIFunction that the agent can call aut
 Same as previous lessons.
 
 ### Step 2: Initialize Plugins
-Create instances of the plugins that provide the functionality:
-- TimeInformationPlugin for time-related queries
-- StockDataPlugin with StocksService for market data
+Set up TimeInformationPlugin and StockDataPlugin for real-time data access.
 
-### Step 3: Create Tools Array
-Convert plugin methods to AIFunction objects using AIFunctionFactory.Create():
-- Time function from timePlugin.GetCurrentUtcTime
-- Stock price function from stockDataPlugin.GetStockPrice
-- Historical data function from stockDataPlugin.GetStockPriceForDate
+### Step 3: Create AI Functions
+Convert plugin methods to AIFunction objects using AIFunctionFactory.Create().
 
-### Step 4: Create Agent with Tools
-Create ChatClientAgent and pass the tools array to enable function calling.
+### Step 4: Create Specialized System Instructions
+Write detailed instructions that define:
+- The agent's role as a stock sentiment analyzer
+- Specific rules for analysis and scoring
+- Required response format and structure
+- Professional behavior expectations
 
-### Step 5: Create Thread
+### Step 5: Create Stock Sentiment Agent
+Use ChatClientAgent with the specialized instructions and tools.
+
+### Step 6: Create Thread
 Same as previous lessons.
 
-### Step 6: Use Agent with Function Calling
-The agent will automatically call functions when needed to answer user questions.
+### Step 7: Process User Messages
+Use agent.RunAsync() with proper error handling for robust operation.
 
-## Testing Function Calling
+## Testing the Agent
 
-Try these queries to see automatic function calling:
+### Stock Analysis Queries
+- "MSFT" - Get sentiment analysis for Microsoft
+- "Analyze AAPL" - Apple stock sentiment
+- "What's your sentiment on TSLA?" - Tesla analysis
 
-### Time Queries
-- "What time is it?"
-- "What's the current UTC time?"
+### Comparison Queries
+- "Compare MSFT and GOOGL sentiment"
+- "Which is better: AAPL or AMZN?"
 
-### Current Stock Prices
-- "What's the current price of MSFT?"
-- "How is Apple stock doing today?"
-- "Get me Tesla's current stock price"
-
-### Historical Data
-- "What was AAPL's price on 2023-01-15?"
-- "Show me Microsoft's stock price from last month"
-
-### Combined Queries
-- "What time is it and what's Google's current stock price?"
-- "Give me a financial report with current time and top tech stock prices"
+### Market Analysis
+- "What stocks should I buy today?"
+- "Give me sentiment analysis for tech stocks"
 
 ## Expected Behavior
-The agent should automatically determine when function calls are needed and execute them seamlessly within the conversation. You'll see the agent providing real-time, accurate data while maintaining its creative and funny personality.
+The agent should provide consistent, structured analysis with:
+- Clear numerical sentiment scores (1-10)
+- Specific buy/hold/sell recommendations
+- Technical reasoning based on current market data
+- Professional, analytical responses
+- Source attribution for analysis
+
+This specialized agent demonstrates how Agent Framework can be used to create professional-grade AI applications with domain-specific expertise and structured outputs.
