+++ workshop-agent-framework/dotnet/Lessons/Lesson4/Lesson5.csproj
