+++ workshop-agent-framework/dotnet/Lessons/Lesson5/Lesson6.csproj
