+++ workshop-agent-framework/dotnet/Lessons/Lesson5/README.md
@@ -1,94 +1,111 @@
-# Lesson 5: Enhanced Financial Analysis with Web Search
+# Lesson 6: Agent Orchestration with Sequential Workflows
 
 ## Overview
-This lesson demonstrates advanced agent capabilities by adding web search functionality to create a comprehensive financial analysis agent. You'll build an agent that can analyze stocks, market trends, and financial topics using both real-time market data and current web information.
+This lesson introduces agent orchestration using Microsoft Agent Framework workflows. You'll learn to create multiple specialized agents and coordinate them in a sequential workflow to perform complex analysis tasks. This demonstrates how to build sophisticated multi-agent systems for enterprise applications.
 
 ## Learning Objectives
-- Integrate multiple data sources (market data + web search)
-- Create sophisticated agent instructions for complex analysis
-- Handle natural language queries and extract relevant information
-- Provide comprehensive analysis with source attribution
-- Build production-ready financial analysis capabilities
+- Understand agent orchestration and workflow concepts
+- Create multiple specialized agents with distinct roles
+- Use AgentWorkflowBuilder for sequential orchestration
+- Handle multi-agent streaming responses
+- Build production-ready multi-agent systems
 
 ## Key Concepts
-- **Multi-Modal Analysis**: Combining structured data (stock prices) with unstructured data (web search)
-- **Natural Language Processing**: Extracting stock symbols and intent from user queries
-- **Source Attribution**: Tracking and reporting information sources
-- **Comprehensive Analysis**: Providing complete responses with multiple data points
+- **Agent Orchestration**: Coordinating multiple agents to work together
+- **Sequential Workflows**: Agents process information in a specific order
+- **Specialized Agents**: Each agent has a focused, specific responsibility
+- **Workflow-as-Agent**: Converting workflows into single callable agents
+- **Multi-Agent Streaming**: Managing real-time responses from multiple agents
 
-## New Components
+## Architecture
 
-### HostedWebSearchTool
-Provides access to current web information for market sentiment, news, and analysis.
+### Three-Agent Portfolio Analysis System
 
-### Enhanced Agent Instructions
-- Analyze various financial topics (stocks, sectors, trends)
-- Extract stock symbols from natural language
-- Use web search for current market sentiment
-- Provide comprehensive single responses
-- Include detailed source attribution
+1. **Portfolio Research Agent**
+   - Gathers market data for all stocks
+   - Retrieves current prices and web sentiment
+   - Provides comprehensive research reports
+
+2. **Risk Assessment Agent**
+   - Analyzes portfolio composition
+   - Calculates risk scores and diversification metrics
+   - Identifies concentration concerns
+
+3. **Investment Advisor Agent**
+   - Synthesizes research and risk analysis
+   - Provides actionable recommendations
+   - Suggests specific buy/hold/sell actions
 
 ## Steps to Complete
 
 ### Step 1: Initialize Chat Client
 Same as previous lessons.
 
-### Step 2: Initialize Plugins with Web Search
-Add the new HostedWebSearchTool alongside existing plugins:
-- TimeInformationPlugin
-- StockDataPlugin  
-- HostedWebSearchTool (new!)
+### Step 2: Initialize Plugins
+Set up TimeInformationPlugin, StockDataPlugin, and HostedWebSearchTool.
 
 ### Step 3: Create AI Functions
-Convert all plugin methods to AIFunction objects, including the web search tool.
+Convert plugin methods to AIFunction objects.
 
-### Step 4: Create Comprehensive Agent Instructions
-Design system instructions that enable the agent to:
-- Analyze stocks, sectors, and general financial topics
-- Extract stock symbols from natural language queries
-- Use web search for current market news and sentiment
-- Apply 1-10 sentiment scoring for stock analysis
-- Provide complete analysis in single responses
-- Include detailed source attribution
+### Step 4: Create Portfolio Research Agent
+Design instructions for comprehensive market data gathering:
+- Stock price collection
+- Web sentiment analysis
+- Structured research reporting
 
-### Step 5: Create Financial Analysis Agent
-Use ChatClientAgent with all tools (time, stock data, web search) and comprehensive instructions.
+### Step 5: Create Risk Assessment Agent
+Design instructions for portfolio risk analysis:
+- Sector concentration analysis
+- Risk scoring (1-10 scale)
+- Diversification assessment
 
-### Step 6: Create Thread
-Same as previous lessons.
+### Step 6: Create Investment Advisor Agent
+Design instructions for recommendation synthesis:
+- Portfolio health scoring
+- Buy/hold/sell recommendations
+- Rebalancing suggestions
 
-### Step 7: Process Complex Queries
-The agent will automatically orchestrate multiple data sources to provide comprehensive analysis.
+### Step 7: Build Sequential Workflow
+Use `AgentWorkflowBuilder.BuildSequential()` to create an orchestrated workflow and convert it to a single callable agent.
 
-## Testing the Enhanced Agent
+### Step 8: Stream Multi-Agent Responses
+Implement streaming that shows which agent is responding and displays their output in real-time.
 
-### Company Analysis (Natural Language)
-- "What do you think about Microsoft?" (should extract MSFT)
-- "How is Apple performing?" (should extract AAPL)
-- "Tell me about Tesla's prospects" (should extract TSLA)
+## Workflow Pattern
 
-### Sector Analysis
-- "How is the tech sector performing?"
-- "Should I invest in renewable energy stocks?"
-- "What's happening in the banking industry?"
+```csharp
+// Create sequential workflow
+AIAgent workflowAgent = await AgentWorkflowBuilder.BuildSequential([
+    researchAgent,
+    riskAgent, 
+    advisorAgent
+]).AsAgentAsync();
 
-### Market Trends
-- "What are the current market trends?"
-- "How is the overall economy doing?"
-- "What are analysts saying about inflation?"
+// Stream responses with agent identification
+await foreach (var update in workflowAgent.RunStreamingAsync(input))
+{
+    // Track which agent is responding
+    // Display formatted output
+}
+```
 
-### Investment Strategy
-- "What should I invest in right now?"
-- "Is this a good time to buy stocks?"
-- "How should I diversify my portfolio?"
+## Testing the Multi-Agent System
+
+### Portfolio Examples
+- "MSFT, AAPL, TSLA, NVDA" (tech-heavy portfolio)
+- "MSFT, JPM, PG, XOM, VZ" (diversified portfolio)
+- "TSLA, RIVN, NIO, LCID" (EV-focused portfolio)
+
+### Expected Flow
+1. **Research Agent**: Gathers data on each stock
+2. **Risk Agent**: Analyzes the portfolio composition
+3. **Advisor Agent**: Provides final recommendations
+
+## Advanced Features
+- **Agent Identification**: Track which agent is responding
+- **Formatted Output**: Professional presentation of results
+- **Error Handling**: Robust operation across multiple agents
+- **Streaming Coordination**: Real-time multi-agent responses
 
 ## Expected Behavior
-The agent should:
-- Automatically determine when to use web search vs. stock data
-- Extract stock symbols from natural language queries
-- Provide comprehensive analysis combining multiple sources
-- Include a dedicated "Sources" section with specific references
-- Give complete responses without requesting more information
-- Apply appropriate analysis frameworks based on query type
-
-This lesson demonstrates how Agent Framework can orchestrate multiple capabilities to create sophisticated, production-ready AI applications.
+You should see a clear progression through three distinct analysis phases, with each agent building upon the previous agent's work to provide comprehensive portfolio analysis. The system demonstrates how complex tasks can be broken down into specialized components that work together seamlessly.
