--- workshop-agent-framework/dotnet/Solutions/Lesson2/Lesson2.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-    </PropertyGroup>
-    
-  <ItemGroup>
-    <ProjectReference Include="..\..\Core.Utilities\Core.Utilities.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.AI" />
-    <PackageReference Include="Microsoft.Agents.AI" />
-    <PackageReference Include="Azure.AI.OpenAI" />
-  </ItemGroup>    <ItemGroup>
-        <None Update="appsettings.json">
-            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-        </None>
-    </ItemGroup>
-</Project>
