+++ workshop-agent-framework/dotnet/Solutions/Lesson2/Lesson3.csproj
