+++ workshop-agent-framework/dotnet/Solutions/Lesson2/Program.cs
@@ -1,24 +1,40 @@
 using Core.Utilities.Config;
+using Core.Utilities.Plugins;
+using Core.Utilities.Services;
 using Microsoft.Extensions.AI;
 using Microsoft.Agents.AI;
 
-// Step 1 - Initialize the chat client using Microsoft Agent Framework
+// Initialize the chat client with Agent Framework  
 IChatClient chatClient = AgentFrameworkProvider.CreateChatClientWithApiKey();
 
-// Step 2 - Create financial advisor agent with conversation context
+// Initialize plugins
+TimeInformationPlugin timePlugin = new();
+HttpClient httpClient = new();
+StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
+
+// Create tools for the agent
+var tools = new AIFunction[]
+{
+    AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime),
+    AIFunctionFactory.Create(stockDataPlugin.GetStockPrice),
+    AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate)
+};
+
+// Create financial advisor agent with function calling capabilities
 string systemInstructions = "You are a friendly financial advisor that only emits financial advice in a creative and funny tone";
 
 ChatClientAgent agent = new(
     chatClient,
     instructions: systemInstructions,
     name: "FinancialAdvisor",
-    description: "A friendly financial advisor that maintains conversation context"
+    description: "A friendly financial advisor with access to time and stock data",
+    tools: tools
 );
 
-// Create a thread for conversation with history
+// Create a thread for conversation
 AgentThread thread = agent.GetNewThread();
 
-// Execute program.
+// Execute program
 const string terminationPhrase = "quit";
 string? userInput;
 do
@@ -37,7 +53,7 @@ do
     {
         Console.Write("Assistant > ");
         
-        // Step 3 & 4 - Use agent to respond with conversation history maintained automatically
+        // Use agent with automatic function calling
         var response = await agent.RunAsync(userInput, thread);
         Console.WriteLine(response);
     }
