+++ workshop-agent-framework/dotnet/Solutions/Lesson3/Lesson4.csproj
