+++ workshop-agent-framework/dotnet/Solutions/Lesson3/Program.cs
@@ -1,8 +1,9 @@
 using Core.Utilities.Config;
 using Core.Utilities.Plugins;
 using Core.Utilities.Services;
-using Microsoft.Extensions.AI;
+using Core.Utilities.Extensions;
 using Microsoft.Agents.AI;
+using Microsoft.Extensions.AI;
 
 // Initialize the chat client with Agent Framework  
 IChatClient chatClient = AgentFrameworkProvider.CreateChatClientWithApiKey();
@@ -12,31 +13,50 @@ TimeInformationPlugin timePlugin = new();
 HttpClient httpClient = new();
 StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
 
-// Create tools for the agent
-var tools = new AIFunction[]
-{
-    AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime),
-    AIFunctionFactory.Create(stockDataPlugin.GetStockPrice),
-    AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate)
-};
+// Create AI Functions from plugins
+var timeTool = AIFunctionFactory.Create(timePlugin.GetCurrentUtcTime);
+var stockPriceTool = AIFunctionFactory.Create(stockDataPlugin.GetStockPrice);
+var stockPriceDateTool = AIFunctionFactory.Create(stockDataPlugin.GetStockPriceForDate);
 
-// Create financial advisor agent with function calling capabilities
-string systemInstructions = "You are a friendly financial advisor that only emits financial advice in a creative and funny tone";
+// Stock Sentiment Agent system instructions - defines the agent's behavior and rules
+string stockSentimentAgentInstructions = """
+    You are a Stock Sentiment Agent. Your responsibility is to find the stock sentiment for a given Stock.
 
-ChatClientAgent agent = new(
+    RULES:
+    - Use stock sentiment scale from 1 to 10 where stock sentiment is 1 for sell and 10 for buy.
+    - Provide the rating in your response and a recommendation to buy, hold or sell.
+    - Include the reasoning behind your recommendation.
+    - Include the source of the sentiment in your response.
+    - Focus on technical analysis based on stock price data and general market knowledge.
+    """;
+
+// Create the Stock Sentiment Agent using ChatClientAgent
+ChatClientAgent stockSentimentAgent = new(
     chatClient,
-    instructions: systemInstructions,
-    name: "FinancialAdvisor",
-    description: "A friendly financial advisor with access to time and stock data",
-    tools: tools
+    instructions: stockSentimentAgentInstructions,
+    name: "StockSentimentAgent",
+    description: "An intelligent agent that analyzes stock sentiment using market data",
+    tools: [
+        timeTool,
+        stockPriceTool, 
+        stockPriceDateTool
+    ]
 );
 
 // Create a thread for conversation
-AgentThread thread = agent.GetNewThread();
+AgentThread thread = stockSentimentAgent.GetNewThread();
 
 // Execute program
 const string terminationPhrase = "quit";
 string? userInput;
+
+Console.WriteLine("=== Stock Sentiment Agent with Microsoft Agent Framework ===");
+Console.WriteLine("This agent analyzes stock sentiment using current market data and technical analysis.");
+Console.WriteLine("Enter a stock symbol (e.g., 'MSFT', 'AAPL') or ask questions about stocks.");
+Console.WriteLine("Type 'quit' to exit.");
+Console.WriteLine("===============================================================");
+Console.WriteLine();
+
 do
 {
     Console.Write("User > ");
@@ -53,9 +73,34 @@ do
     {
         Console.Write("Assistant > ");
         
-        // Use agent with automatic function calling
-        var response = await agent.RunAsync(userInput, thread);
-        Console.WriteLine(response);
+        try
+        {
+            // Use the agent to process the user message
+            var response = await stockSentimentAgent.RunAsync(userInput, thread);
+            
+            // Extract and display the response content
+            if (response?.Messages?.Any() == true)
+            {
+                var lastMessage = response.Messages.Last();
+                Console.WriteLine(lastMessage.Text ?? "No response generated.");
+            }
+            else
+            {
+                Console.WriteLine("No response generated from the agent.");
+            }
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"Error processing request: {ex.Message}");
+            if (ex.InnerException != null)
+            {
+                Console.WriteLine($"Inner exception: {ex.InnerException.Message}");
+            }
+        }
+        
+        Console.WriteLine();
     }
 }
 while (userInput != terminationPhrase);
+
+Console.WriteLine("Thank you for using the Stock Sentiment Agent!");
