+++ workshop-agent-framework/dotnet/Solutions/Lesson5/Lesson6.csproj
